Creating Visual Studio 2022 solution for GLQuake
There are a few minor changes included. GetVersionEx is deprecated, so I removed the call to it in Sys_Init. A few define gaurds that were part of the previous build process have been removed. A segfault in console.c involving vsprintf has been fixed. There were several places where errno was used as a variable name, which was making the compiler through a fit. I renamed those to err. At this point textures are not loading correctly.